fix(apps/k8s-gateway): break DNS req loop using TLS
AdGuard is quite eager in how it handles DNS queries - it intercepts and
responds to all plaintext queries, regardless of the target server for
the query. Since there is a split-horizon DNS setup in the AdGuard
settings, it forwards all queries to k8s-gateway, even when k8s-gateway
has no response to a query and tried to get the response from the
upstream. By forcing `forward` plugin to use TLS, AdGuard cannot
intercept and respond to a query, and so k8s-gateway can respond to
queries it doesn't know about by forwarding an upstream response.